total refactoring autocomplete to respect Tony and static analizators :) p.s. I'm fix last leaks

+++ includes/autocomplete.h
@@ -32,8 +32,8 @@
 
 char			**autocomplete(t_line *line_info, int pos);
 extern char		*g_built_in_lists[];
-char			*atcml_get_pattern(char *buf, int pos);
-char			*atcml_get_rel_dir(char *buf, int pos);
+char			*atcml_get_pattern(char *buf, size_t pos);
+char			*atcml_get_rel_dir(char *buf, size_t pos);
 int				check_env_a_case(char c, char cb, int pos);
 int				check_oth_a_case(char c, int pos);
 int				get_autocomplite_type(t_line *line_info,
@@ -56,8 +56,8 @@ char			**get_autocomplite_files_dir_mas(char *str, char **res, int *c);
 char			**get_mas_other_autocompile(char *str);
 char			**get_mas_of_suggetions(char *word, int type);
 char			**sugg_delete_repeats(char **str);
-int				sugg_check_repeats_in_all_mass(char **str, int pos, char *c);
-char			**sugg_free_and_set_one(char **str, int pos);
+int				sugg_check_repeats_in_all_mass(char **str, size_t pos, char *c);
+char			**sugg_free_and_set_one(char **str, size_t pos);
 char			**sugg_get_common_repeat(char **str, char *sugg_word);
 
 #endif
